Update to fix build issues

+++ applications/stlp/src/usb/adaptive_rate_callback.h
@@ -7,7 +7,7 @@
 #include <stdbool.h>
 
 
-float determine_USB_audio_rate(uint32_t timestamp,
+uint32_t determine_USB_audio_rate(uint32_t timestamp,
                                     uint32_t data_length,
                                     uint32_t direction,
                                     bool update);
